test: add gen_elders_info helpers
